+++ shopify/packages/hybrid-cart-api/src/effect/schema.ts
@@ -8,7 +8,7 @@ export const BaseAttributes = Schema.Record({
     Schema.Null,
     Schema.String,
     Schema.Boolean,
-    Schema.Number,
+    Schema.Number
   ),
 });
 
@@ -142,6 +142,25 @@ export const LineLevelDiscountAllocation = Schema.Struct({
   discount_application: DiscountApplication,
 });
 
+const CartLevelDiscountApplicationUpdate = Schema.transform(
+  Schema.Struct({
+    discount_application: DiscountApplication,
+  }),
+  DiscountApplication,
+  {
+    decode: (application) => application.discount_application,
+    encode: (discountApplication) => ({
+      discount_application: discountApplication,
+    }),
+    strict: false,
+  }
+);
+
+export const CartLevelDiscountApplication = Schema.Union(
+  DiscountApplication,
+  CartLevelDiscountApplicationUpdate
+);
+
 export type UnitPriceMeasurement = Schema.Schema.Type<
   typeof UnitPriceMeasurement
 >;
@@ -198,7 +217,7 @@ export const LineItem = Schema.Struct({
   unit_price_measurement: Schema.optional(UnitPriceMeasurement),
   selling_plan_allocation: Schema.optionalWith(
     Schema.NullOr(SellingPlanAllocation),
-    { default: () => null },
+    { default: () => null }
   ),
 });
 
@@ -218,7 +237,7 @@ export const Cart = Schema.Struct({
   requires_shipping: Schema.Boolean,
   currency: CurrencyCode,
   items_subtotal_price: Schema.Number,
-  cart_level_discount_applications: Schema.Array(DiscountApplication),
+  cart_level_discount_applications: Schema.Array(CartLevelDiscountApplication),
 });
 
 export const makeCartSchema = (sections?: string) => {
@@ -227,7 +246,7 @@ export const makeCartSchema = (sections?: string) => {
       ...Cart.fields,
       sections: Schema.optionalWith(
         Schema.NullOr(Ajax.Sections.makeResponseSchema(sections)),
-        { default: () => null },
+        { default: () => null }
       ),
     });
   }
@@ -268,7 +287,7 @@ export const UpdateItemRecordInput = Schema.Record({
 export type CartUpdateInput = Schema.Schema.Encoded<typeof CartUpdateInput>;
 export const CartUpdateInput = Schema.Struct({
   updates: Schema.optional(
-    Schema.Union(UpdateItemRecordInput, Schema.Array(Schema.Number)),
+    Schema.Union(UpdateItemRecordInput, Schema.Array(Schema.Number))
   ),
   note: Schema.optional(Schema.NullOr(Schema.String)),
   attributes: Schema.optional(Attributes),
@@ -294,7 +313,7 @@ export const CartChangeItemOptionalInput = Schema.Struct({
     Schema.Record({
       key: Schema.String,
       value: Schema.NullOr(Schema.String),
-    }),
+    })
   ),
   selling_plan: Schema.optional(Schema.NullOr(Resource.ID)),
 }).pipe(BaseInput);
@@ -324,13 +343,13 @@ export const CartChangeInput = Schema.Union(
   Schema.extend(CartChangeItemOptionalInput)(
     Schema.Struct({
       id: Schema.String,
-    }),
+    })
   ),
   Schema.extend(CartChangeItemOptionalInput)(
     Schema.Struct({
       line: Schema.Number,
-    }),
-  ),
+    })
+  )
 );
 
 export type CartChangeOutput = Schema.Schema.Type<typeof CartChangeOutput>;
@@ -350,5 +369,5 @@ export type CartUpdateDiscountsInput = Schema.Schema.Encoded<
   typeof CartUpdateDiscountsInput
 >;
 export const CartUpdateDiscountsInput = Schema.mutable(
-  Schema.Array(Schema.NonEmptyString),
+  Schema.Array(Schema.NonEmptyString)
 );