+++ shopify/packages/hybrid-cart/src/mutations/updateDiscounts.ts
@@ -17,8 +17,12 @@ export const mutation = createMutation(
 
       return response.data;
     },
-    onSettled() {
+    onError() {
       Cart.invalidate();
+    },
+    onSuccess(cart) {
+      if (!cart) return Cart.invalidate();
+      Cart.set(() => cart);
     },
   }),
   () => client,