+++ shopify/packages/hybrid-cart-components/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/shopify-hybrid-cart-components
 
+## 0.2.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/shopify-hybrid-cart@0.1.26
+
 ## 0.2.15
 
 ### Patch Changes